add AdminWalletService
